Experiment Autodiff in polar
Previously, D(x) is provided as contant tensor, so actually \grad D(x) not computed

+++ config.py
@@ -2,7 +2,7 @@ import os
 import tensorflow as tf
 
 # Set data type
-DTYPE='float64'
+DTYPE='float32'
 tf.keras.backend.set_floatx(DTYPE)
